Tag: Fix codegen when using 2.4.0 swagger-codegen
After upgrading to swagger-codegen 2.4.0, it appears that it no longer
generates models that are missing a type and instead require we include
type: object. Additionally, it's having trouble generating arrays of
inlined objects, resulting in things like List<Object> instead of it
being a List<MyModel>. The workaround is to pull out every inline object
of an array into a definition. The resulting codegen is identical to
what was produced by swagger-codegen 2.3.1.

We're also using this opportunity to rename some of the selection API
models to be more consistent, which does result in a codegen change.