We will support only multi function jobs.
Changed the JobModel to support only multi-function requests. Fun has
to be a list of strings. We don't support 'string' case anymore because
we want to be able to use DyanmicExpressions to query for jobs data.
They don't work if the JobModel is not "well" defined(inheritances are not supported)
So to be able to use the projection over the models in
backend we will support only arrays. And the API looks much better now.

Signed-off-by: Cristian Hotea <cristian.hotea@ni.com>