Merged PR 27347: [API Review] Enumerate Test Monitor V2 API operations in the service's base route and /{version} route
Adds a V2Operations map in addition to the V1Operations one.

Dynamically generated the operations maps instead of generating a class for each one. This makes it easier to rev API versions and add operations in the future, since they aren't statically defined in the OpenAPI doc. The responses themselves are the kept the same.