FileIngestion: add support for synchronous metadata/data ingestion
Modify upload-files endpoints to support synchronous upload of data and
metadata. This endpoint used to allow multi-file upload on the same
request. This change removes that capability since it was a feature not
used, and not well structued (only the last FID was returned). This
change adds a new formData part to upload a JSON dictionary of metadata
key/value pairs. When uploading metadata, the file part needs to have a
fieldName of "file" and the metadata part a fieldName of "metadata". The
service will continue to accept single part requests, without a
meaningul fieldName.

Signed-off-by: Alejandro Del Castillo <alejandro.delcastillo@ni.com>