Merged PR 29695: Remove an incorrect enum value from the StepField enum
Removed `ID` from the `StepField` enum in the swagger doc and MIDL. `ID` isn't a field on steps. It is a model field, but it's not user visible and there's no benefit in sorting by it.

This doesn't break backwards compatibility - using the enum value currently causes the service to throw an exception. It should instead return a 400, because the value is not valid.

Also fixes an exception message that isn't helpful.