Merged PR 30264: RPC-ify Test Monitor update routes
Changes all the PUT `collection` routes in the existing API to a more correct RPC-style POST `update-collection`.

Since update-results and update-steps had V1 compatibility tags, added the old versions to the Deprecated section to ensure backwards compatibility is kept.

Abstracted one of the request bodies into an object, because the generator was producing conflicting classes for it.

Updated the client API to use the V2 version. Ran integration tests against a server with these changes present.