Merged PR 31704: FileIngestion: add metadata versioning
# Justification
Closes race condition during metadatadata updates

# Implementation
Currently, there is no way to prevent two clients from blapping each other metadata since the metadata is not versioned. To prevent that from happening, added a metaDataRevision key to the mongo store. The files route will now return that key/value pair. That value is meant to be passed to the update-metadata route on the expectedRevision parameter. For backwards compatibility, the parameter is optional.

# Testing
- Ran all unit/integration tests
- Verified database migration
- Manually ran curl to make sure the files route returns metaDataRevision
- Manually ran curl to make sure metadata is updated under the proper conditions:
     * When not providing expectedRevision
    *  When providing expectedRevision at the correct revision
- Manually ran curl to make sure metadata is NOT updated when expectedRevision is provided and is incorrect

# Checklist
- [x] I tested changes to product code in product
- [x] I considered updates to the wiki