Merged PR 34303: Test Monitor OpenAPI: format "iso-date-time" to "date-time"
# Justification
"date-time" is the standard way to specify the format of a DateTime object:
https://swagger.io/specification/

The OpenAPI Generator will generate a Python client that will correctly deserialize to a Python datetime object if the format is "date-time" but not if the format is "iso-date-time". "iso-date-time" is something made up at NI and is not in use anywhere else.