Merged PR 34131: Implementation for query-utilization RPC.
# Justification
We need a query asset utilization route in order to bring data in to a Jupyter Notebook.

There is another review that includes just the midl. Please ignore the midl file.

# Testing
Ran the tests. After the algorithm for computing utilization percentages is in there will be more tests for this.

Related work items: #145681