Merged PR 35093: AssetAvailabilityFacade uses working hours to compute the availability.
# Justification
Computing assets availability must get as input only dates and compute the availability only during working hours.

# Implementation
Working hours are set on the service policy. The AssetAvailabilityFacade get the ServicePolicy in the constructor to be able obtain the working hours when availability needs to be computed. A request interval is splitting in sub intervals with respect to working hours. All sub intervals have the same name as the interval that was split. This helps the algorithm to compute the result for subintervals in a single place. So if the request was for a week, the availability for each day is computed in the same Duration object, having as key the name of the week.

I tried to do minimal changes in the algorithm and facade.

# Testing
Run unit and integration tests. Added some tests that have 9-17 working hours.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki