Merged PR 35353: Added exportPdf and exportHtml as options in the v1 and v2 query routes
# Justification
When a notebook execution is asked to be generated, and then exported as a pdf, the user may not have the proper libraries installed on the server for a successful export. Adding the export options to the queries allows the server to determine whether it can export successfully (user has installed wkhtmltopdf and the exe's path is in the PATH variables). I have added it to both v1 and v2 versions.

# Implementation
TODO: Add exportPdf and exportHtml to the v1Options and v2Options enums.

# Testing
TODO: Manually test to see if the operations route returns an accurate reflection of the systems capabilities to export as a pdf or to report false for no privilege ( no wkhtmltopdf installed or path to exe is not in PATH).
TODO: All automated testing

# Checklist
- [ ] I tested changes to product code in product
- [x] I considered updates to the wiki