Merged PR 41313: Query System Utilization
# Justification
To display the utilization grouped by system, we cannot use the the query asset utilization request because it cannot be configured to give the result a user wants. To I created a new AMQP RPC and backend algorithm for this.

# Implementation
The asset utilization algorithm groups utilizations by asset identifier. I created a base class for the utilization algorithm and extracted that filter into a subclass. Then created a new SystemUtilizationAlgorithm. The only difference between this and the AssetUtilizationAlgorithm is the filtering method, that for systems filter by minion ID. The new algorithm is called from a new RPC, QuerySystemUtilization. Added tests for these.
Obtain the name of the managed systems in the backend and put that on the reply for each UtilizatonWithPercentage object. This is needed so that notebooks don't need to use the sysmgmt HTTP API to obtain that (it is unknown how to authenticate from the notebook).

# Testing
Ran all APM unit and integration tests.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki