Merged PR 44254: Add discoverSystems operation on SystemsManagement HTTP API
# Justification
BUG 216244
CAR 745582
We need a way to tell the HTTP clients that the discovery Service isn't installed. We need this to hide the Discovered widget in SystemsManagement Web UI and to prevent polling the routes implemented in this service.

# Implementation
Since the ServiceManager doesn't have an HTTP API to tell us which service is running or not we have to expose this behavior through the the Systems Management HTTP API. So the approach is to add a new Operation that refers to discovering/registering targets so this would be an addition to the current API. Besides this the description of the manageSystems operation will be modified to refer only to managing keys and removing systems. Theoretically this is a breaking API change since we are redefining what manageSystems means. However the manageSystems available logic will be computed the same way(based on http write privilege from the request) so we will not return a different value after this change. So from the behavior point of view, it will not be a breaking change. We will not impact anybody with this change.

# Testing
Tested along with the UI changes.

# Checklist
- [x] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #216244