Merged PR 44637: Install swagger doc containing the V2 tag historian API
# Justification
This change removes the old tag historian swagger doc, as we're dropping support for that API, and replaces it with the doc describing the V2 API. This also means that we can remove the cloud doc. I'm also removing the flush-values route from the doc, as the route does nothing in the elixir service. I'm making this change to AzDO, rather than GitHub, as the new service will not ship in SLS until 19.6 is released.

# Implementation
I replaced the contents of nitaghistorian.yml with the contents of nitaghistorian-cloud.yml. I removed "Cloud" from the description, and removed the flush-values route.

# Testing
 I pasted the document into editor.swagger.io and verified that there were no errors.

# Checklist
~~- [ ] I tested changes to product code in product~~
~~- [ ] I considered updates to the wiki~~

Related work items: #228973, #228979