Merged PR 47406: Tag queries
See API Review: !45520.  I will pull the changes in that branch into this PR before submitting anything.

# Justification
I'm adding a new route to the Tag HTTP service that supports querying tags using Dynamic LINQ, so that Dashboards can get a sorted list of tags and the tag service is consistent with other APIs' query routes.

# Implementation
To perform the query, I'm pulling tags and values into memory because there is not a way to run a LINQ query directly on Redis. Because of this, some functionality is not supported, e.g. using projection: new(current.value as value) when there are null values in the tag set.

I've decided to use the Systems Manager way of doing projections because the tag with values object is better filtered using nested statements.

Ordering by value always groups tags by type and returns the types in the following order: numbers, strings, booleans, datetimes, and nulls of any type. This was a decision made by Mark and reflective of popularity of our different tag types. Group order is not dependent on `descending`.

The continuation token is a base64-encoded skip. When the entire set of tags matching the query has been returned, the continuation token will be null.

# Testing
The modified Tag Service is available on http://yeti2/ (admin/skyline). I also added python tests to the ATS.

# Checklist
- [x] I tested changes to product code in product
- [x] I considered updates to the wiki - I updated [Skyline Error Codes](https://nio365.sharepoint.com/sites/skyline/SitePages/Skyline%20Error%20Codes.aspx)

Related work items: #216489