Merged PR 50870: Fix 936539 and 936541
# Justification
Bug fixes.

# Implementation
I've removed some periods and added `Take` to the `TagQuery` that is created when the POST body is empty. Now when you send a POST query-tags-with-values with no body, it will return all your tags instead of an empty array and continuation token.

# Testing
Verified on http://yeti2/ (admin/skyline)
Ran ATS for query-tags-with-values locally

# Checklist
- [x] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #936539, #936541