Merged PR 50973: Make query utilization history order by parameter work as defined in Swagger
# Justification
https://ni.visualstudio.com/DevCentral/_workitems/edit/936298
Order by does not work in query utilization history the way is defined in the swagger doc.

# Implementation
Changed to use transform from the strings defined in the swagger doc (eg: "ASSET_IDENTIFIER") to the expression to sort on (a => a.AssetIdentifier).

# Testing
Ran all unit and automated tests.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #936298