Merged PR 51698: Add "Model" suffix to some models that did not have it
# Justification
Some models in the Swagger have "Model" suffix and some don't. They need to have the same naming schema that would help in better identifying which generated class belong to which layer.

# Implementation
Add "Model" suffix to some models in the Swagger file. Don't add "Model" suffix to objects that have string type or enum.

# Testing
N/A

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki