Merged PR 54271: Implement delete assets route
# Justification
Implement Swagger generated interface

# Implementation
Implement delete assets method. Added a result object to be returned by the assets facade. This object is needed to be able to create the errors to be returned by the route. The object contains the ids of the deleted assets, the invalid ids and the not existing asset ids.

# Testing
Ran all unit tests.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #947903