Merged PR 56783: query-data returns partial results if there is a mismatch between x and y
# Justification
The need is to return data for the groups where we can return data, but return errors for the groups where there are errors. Also, possibly, return data AND errors for some groups depending on the error. For example: we truncated x or y because they didn’t match, so we’ll return you ‘something’ but it is lossy.

# Implementation
Add partial error return to query-data. Returning errors in error.

# Testing
added and modified existing tests. Tested in product.

# Checklist
- [x] I tested changes to product code in product
- [ ] I considered updates to the wiki