Merged PR 57169: [S] [HotFix] Removed 'required' attribute from some utilization response properties which can be empty.
# Justification
When we don't have utilization stored for some assets, we pad the response with utilization entries with 0 percent and no minion ID/utilization category.
This was an issue as all the utilization properties were marked as 'required' and JSON serialization of the response would fail as they would be generated in C# with the Required=Always attribute.

# Checklist
- [X] I tested changes to product code in product
- [ ] I considered updates to the wiki