Merged PR 58209: Obtain user name and id differently until we implement the authentication
# Justification
Until we implement authentication properly, we have to obtain the user name and id differently as we do now. One integration test fails because of this.

# Implementation
Update the Swagger document to get the user name and id passed to the request object and get it from there.

# Testing
Ran all integration tests and the one that failed is passing.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki