Merged PR 58087: Update NotFound error example and code
# Justification
Update description of the the NotFound error message and update code to return the asset not found error for multiple routes. All routes use the same generated NotFound error model.

# Implementation
Changed yml documentation with the actual message and update code to return the id of the resource.

# Testing
Ran unit tests.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #962498