Merged PR 58067: Updated delete-assets route implementation to return Not Found and updated the partial response
Added not found case to the web service facade delete assets method when no assets match the asset identifiers in the request.

Updated the route to return a DeleteAssetsResponse which contains the list of identifiers for assets which were deleted, the list of identifiers which failed to be deleted and a list of errors.