Merged PR 58119: Implemented the end multiple utilizations http route.
# Justification
We need to enable the users to update multiple utilization entries with the end timestamp over HTTP.

# Implementation
Implemented the web service logic and added a new method in the UtilizationFacade. This new method allows multiple utilizations to have the end timestamp updated. Up until now, updating the end timestamp was possible for one utilization at a time.

# Testing
Added tests for the new method from the UtilizationFacade and for the new method in the WebServiceFacade.

# Checklist
- [x] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #961050