Merged PR 58395: Update query assets example to work with no changes
# Justification
The example filter for query-assets was too specific: filter assets with a specific name and take 0. It should be updated to a more generic filter so that no matter what you have in your database to get something, and take all (-1).

# Implementation
Add "example" next to "x-example" so that we have take: -1. Update the asset filter to obtain all system controllers.

# Testing
Checked yml in editor.swagger.io

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #962510