Merged PR 58397: Update Destination description and example for query location and calibration
# Justification
The provided example has JSON/INLINE which is an invalid combination. Change the example to CSV, DOWNLOAD.
Reusing response Destination causes some weird documentation when viewing it on the query-locations route: "INLINE (default) returns the list of assets...". We should rename to "resources"

# Implementation
Changed "assets" to "resources" in the documentation. Provide a working example to the query (override the default one).

# Testing
View Swagger in the editor and run the example.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #962517, #962518