Merged PR 58541: Add optional ID part to file upload request
# Justification
The File AMQP API supports a client pre-determining the ID when uploading a file. This is used to solve an issue in the store & forward where file IDs are assigned after upload, but these IDs must also be attached to test results. This is extremely difficult to solve using other means, because files and test results are different services, and may not both be present at the same time.

# Implementation
The multipart/form-data request to upload a file now supports an optional third part, `id`, which may contain a UTF8-encoded file ID. If this is supplied, the service will use this ID for the new file, rather than generating one. If this provided ID is invalid, the service will simply generate its own, rather than reject the request.

# Testing
Uploaded a file using postman, specifying the ID. Verified that the file was created in FIS with the expected ID.

# Checklist
- [X] I tested changes to product code in product
- [X] I considered updates to the wiki