Merged PR 58819: Added utilization ids arrays for failed and succeeded end utilization updates. Also 404 -> 200.
# Justification
Added utilization ids arrays for failed and succeeded end utilization updates. Also 404 -> 200.

# Implementation
The utilization facade returns a summary of the succeeded and failed operations and the web service facade builds the errors.

# Testing
Updated the existing tests. Since the web service deals with the error codes now, a new test was added to ensure the proper error codes are returned.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #964584