Merged PR 58790: Do not return 404 when the URL is found
# Justification
In SystemLink, we return 404 only if the URL is not found. If the URL is found, but some/all resources in the request body are not found, partial success or bad request or empty response is returned, depending on the actual route.

# Implementation
Changed several routes to not return 404 any more. Update the Swagger and the web service facade.
Routes covered:
- compare availability
- delete assets
- query assets
- query asset utilization
- read asset utilization

# Testing
Ran all unit tests.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #966272, #966308, #966348, #966349, #966874