Merged PR 59233: Update peristance model for Assets to contain AssetType and DiscoveryType
# Justification
In accordance with the design doc here: !52380

# Implementation
This is the follow-up change to !57335 in the series of changes to split the AssetClass into an AssetType and a DiscoveryType for assets.

Here's the diagram to illustrate the plan:
https://ni.visualstudio.com/94b22d7b-ad7b-4f5e-88f0-867910f91c94/_apis/git/repositories/8159353b-4403-4eed-9a32-0c2ee30abe3d/Items?path=%2FSpecs%2FAsset+Management%2F[20.0]+Asset+Salt+Components+over+HTTP%2FAssetClassChanges.png&versionDescriptor[versionOptions]=0&versionDescriptor[versionType]=2&versionDescriptor[version]=d2ccbf15a49bec3a97981df8aee69b90bd7c2019&download=false&resolveLfs=true&%24format=octetStream&api-version=5.0-preview.1

This change updates the persistence model to contain the AssetType and DiscoveryType properties, while removing the AssetClass and IsSystemController properties.

The AssetClass property is removed throughout the layers, as it was a half-baked way of specifying both discovery type and asset type, and we're moving away from that. The IsSystemController is currently exposed on the query-assets route, so we're going to continue supporting it, but this will just be a computed property on the DomainModel alone.

This also requires the following changes:
1. Create a migration to create the new properties and drop the existing ones (see MigrationManager.cs)
2. Preprocess the filter `"IsSystemController = True"` to be converted into an `"AssetType = \"SYSTEM\""` filter (see AssetsFacade.cs)

The migration happened in accordance with the table bellow:
| IsSystemController|AssetClass|**DiscoveryType**|**AssetType**|
|---|---|---|---|
|false|Unspecified|Manual|Generic|
|false|Discovered|_Automatic_|Generic|
|false|Asset|Manual|Generic|
|false|Dut|Manual|DeviceUnderTest|
|false|Fixture|Manual|Fixture|
|true|Unspecified|Manual|_System_|
|true|Discovered|_Automatic_|_System_|
|true|Asset|Manual|_System_|
|true|Dut|Manual|_System_|
|true|Fixture|Manual|_System_|

# Testing
Updated existing tests to verify the new properties.

# Checklist
- [x] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #961048