Merged PR 59824: AM HTTP Route Update - Start Utilization should return partial success instead of 404 and the failed and successful asset identifications.
# Justification
Start Utilization should return partial success instead of 404 and the failed and successful asset identifications.

# Implementation
The utilization facade deals with starting the utilization and returning a class that contains a summary of the failed and succeeded operations.
The webservice then builds the error message and returns it.

# Testing
Updated all the tests to validate the new specification.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #964583, #966877