Merged PR 60729: Add numeric error codes to TDM Reader and increase request timeout
# Justification
The LabVIEW API relies on numeric error codes to populate the error wire. With the switch to OpenAPIGenerator instead of a pass-through web service, the TDM Reader web service stopped passing numeric error codes back to clients. Additionally, the request timeout when from 60 seconds in the pass-through to the AMQP web service default of 30 seconds.

# Implementation
Add numeric error codes to all TDM Reader and file fetcher errors. Note that the FileIdIncorrect error was not being used, was not in the error dictionary, and not in the Skyline Error Codes document, so I removed it.

Set request timeout to 60 seconds for all routes. Long term, we should instead add calls to `IHttpRequest.ResetRequestTimeout` to indicate progress is being made. This will be cherry-picked to 20.0.

# Testing
Ran the LabVIEW tests and verified they now all pass. Examined the generated web service INI file and verified the 60 second request timeout is present on all handlers.

# Checklist
- [x] I tested changes to product code in product
- [ ] ~~I considered updates to the wiki~~

Related work items: #972314, #972321