Merged PR 60666: Validate calibration entry before creating or updating asset.
# Justification
Need to validate a calibration entry before using it to create or update an asset. This prevents the user for specifying a wrong next calibration date or invalid recommended interval etc.

# Implementation
Extract the validation done for adding external calibration for 3rd party assets into an extension class (to not pollute the model with ErrorCode and ITraceLogger). Use the validation extension method from the mapping used for create and update assets. Catch the exception at the right level for each case. Unit test.

# Testing
Wrote unit tests.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #973980, #973982