Merged PR 61157: add ignoreXForDecimation, which allows graphs to change x axis and not use sort
# Justification
Sorting + decimation on the graph needs to be able to ignore the synthetic or user-passed x and instead return synthetic indices. Otherwise, sorting has no practical effect on the graph and decimating sorted data is not possible, apart from the sort being lossy.

# Implementation
Add parameter specifying that the x (synthetic or passed) that we use for decimation will be ignored.

# Testing
Added new tests, tested in argon, ran existing tests.

# Checklist
- [x] I tested changes to product code in product
- [ ] I considered updates to the wiki