Merged PR 61260: Updating one or more identification properties return partial update instead of failed
# Justification
Updating the serial number or other identification property fails when the asset identifier is provided. This is ok, but it should be partial update in order to update other properties that were specified. Also, the error must say we cannot update the serial number instead of saying that building the asset identifier failed.

# Implementation
If the id is on the request, use that and don't try to build one from the identification properties. In that case, if the identification properties are different that then stored asset ones, return error/warning that those could not be updated.

# Testing
Added a unit test that fails without this change and passes with this change.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #976106