Merged PR 61970: New route updating multiple system in Systems Management HTTP API
# Justification
We want to modify the metadata (for the moment the workspace) for more system in one call.

# Implementation
Defined a new rpc route of type POST that operate on multiple systems. For this, we use a list of systems ids and a metadata property defined in a patch way. There will be bools that show if the values are set or not for the metadata information.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki