Merged PR 67997: Add Query Assets Availability Route in YAML
# Justification
In order to update our notebooks to use HTTP instead of AMQP we need to start by ensuring that we have similar HTTP routes to those exposed by our AMQP API.

We currently do not have a http route which returns availability for multiple assets during multiple time intervals. A workaround would be to call the compare availability route several times, but this would lead to a significant decrease in performance.

# Implementation
Modified the yaml document and implemented the new method in web service facade. The final implementation of this method will follow in the next PR.

# Testing
Ran code gen, checked the generated models.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #1001821