Merged PR 69983: Remove ni-auth atribute from ping and packages download routes.
# Justification
Anyone should be able to ping the service even without authentication.
The package pool is public so any user with the exact download link of a package should be able to download it. Also the  resources of repository service are feeds, not packages.

# Implementation
removed 'x-ni-auth: true' from /v1/files get route and /v1/ping get route from the yml file.

# Testing
Pinged the service and download a package from a unauthenticated private browser session.

# Checklist
- [x ] I tested changes to product code in product
- [ ] I considered updates to the wiki