Merged PR 74215: Query System Utilization to http
# Justification
We need to expose the system utilization to http so that it can be used in Asset Utilization Notebook Report.

# Implementation
In the yml file I created QuerySystemUtilizationsRequest which represents the request model. Also in yml I created a model that represents the returned type of the result. Although query asset utilization and query  system utilization return an object that contains the same properties and on amqp we could use the same model, on http I had to create a new one for query system utilization ('SystemUtilizationWithPercentageModel')  because on the model returned by asset utilization, the assetIdentifier property is required and this property on query system utilization is always null.

Also in the yml file I entered the / v1 / query-system-utilization route using the created models.

# Testing
Unit test
I tested the route with Postman

# Checklist
- [x] I tested changes to product code in product
- [ ] I considered updates to the wiki

Query System Utilization to http

Related work items: #1001825