Merged PR 78830: Removed assetId from QuerySystemUtilization Response
# Justification
On the query system utilization response was returned an assetId that represents the id of a random asset in the system. This didn't bring any value so I removed the attribute from the message.

# Implementation
Removed assetId from SystemUtilizationWithPercentageResponse from yml file.

# Testing
I tested in the product
I ran unit tests

# Checklist
- [x] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #1047560