Merged PR 79075: Forward ApiKey from asset manager to file ingestion service
# Justification
APM service has several routes that send a query result to the FileService. We must forward the ApiKey when sending the request to store the file. The user must specify the workspace where to store the file.

# Implementation
Update all routes (HTTP and AMQP) that sends data to FileIngestionService to have a new parameter that tells in which workspace to put the file. Forward the ApiKey. When the user does not have permissions to store the file, the error is forwarded back to the client (no change for this, already works).

Enable again the FileIngestion integration tests and update then to have FileIngestionService permissions and use a valid workspace id.

# Testing
Tested local in the product and make all integration tests pass.

# Checklist
- [x] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #1037900