Merged PR 80256: Update HTTP API to include system name and asset name in the query system
# Justification
We moved the AMQP query utilization APIs to HTTP and for some reason we forgot to add some properties that are used by the asset utilization notebook in the group by asset and group by system cases.

# Implementation
Added back the properties that are required by the asset utilization notebook, to not change the notebook to obtain those again.

# Testing
Tested locally.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki

Related work items: #1064880