Merged PR 86066: Include metadata blob on notebook execution records
# Justification
The execution record includes all the notebook data (parameters, path, etc) except the metadata blob.

The V2 notebook output format only stores the version number in the metadata blob, and not in the actual notebook output. Combined with the above, this means that execution records don't actually say which version the results are in - the output blob is present in the execution result without an actual version number.

We also can't simply start adding the version number to notebook output either, since there now exist notebooks in the V2 format which do not record a version number. Therefore, the only non-breaking way to ensure the version number appears in the execution record is to add the metadata blob.

# Implementation
Updates the swagger doc to include the metadata blob. Corresponding implementation changelist is WIP on the perforce side.

# Testing
Rendered this documentation in the Swagger editor.

# Checklist
~~- [ ] I tested changes to product code in product~~
~~- [ ] I considered updates to the wiki~~
- [X] If this PR affects user-visible strings, I added Jenna Jaco as a reviewer and attached screenshots of the strings in the product