Merged PR 88195: Add reinstall flag to the API
# Justification
The python module already supports this but defaults to `False` (which is the expected use-case):
```python
def installed(name, version, reinstall=False, **kwargs):
```
I'm just adding it to the HTTP API too, so that users have the option to erase targets. Still defaults to false, since this is more like an advanced use-case.

__I think it's ok if this new flag is not reflected in the UI (that can be another change). But at least we officially support this in the raw editor and using the HTTP APIs__

# Implementation
Add optional flag to the model

# Testing
- Updated existing tests (default use-case with `False`)
- Added new test for `True` flag

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki
- [ ] If this PR affects user-visible strings, I added Jenna Jaco as a reviewer and attached screenshots of the strings in the product

Related work items: #1108603