Merged PR 86381: Bug 1094146: Refactoring DeleteMultiple to handle errors more accurately
# Justification
The initial issue was that we first delete the file from the database and then try to delete the file. It's better to do this the other way round because it's less likely that we fail to delete the database entry after we deleted the file.
While investigating it turned out that we don't handle errors very smoothly if multiple files couldn't be deleted.

# Implementation
Refactored the strategy how to delete multiple files and handle a variety of common outcomes.

# Testing
Manual testing

Related work items: #1094146