Merged PR 95491: [File Moving Service] Get file path using additional policy action on FileIngestion query route
# Justification
For the FileMoving service to use HTTP for RPC communication, the file path on the server needed to be changed from the AMQP message to accessible over HTTP.

# Implementation
Implemented a new file-paths:Read action in the FileIngestion service.  This is used during the file query and returns file paths only if the API key used in the request has the privilege.

# Testing
Installed the resulting FileIngestion and FileMoving binaries on my server and verified that the FileIngestion service returns the file path when the privilege is present and returns a null path element when the privilege is not present.

# Checklist
- [x] I tested changes to product code in product
- ~~[ ] I considered updates to the wiki~~
- ~~[ ] If this PR affects user-visible strings, I added Jenna Jaco as a reviewer and attached screenshots of the strings in the product~~