Merged PR 117358: Asset Service - Asset route for the query assets availability report.
# Justification
We need to find o way of making the asset reports execute faster.
In this case, the Connection History report will benefit from using a route that returns the data in csv format.

# Implementation
A new route was added to the asset service that returns availability data in the csv format.

Pull Request 116981: Add a csv route for asset availability.
https://ni.visualstudio.com/DevCentral/_git/Skyline/pullrequest/116981

# Testing
Check the PR linked above, it contains the required tests for the route and also some performance benchmarks.

# Checklist
- [x] I tested changes to product code in product
- [x] I considered updates to the wiki
- [x] If this PR affects user-visible strings, I added Jenna Jaco as a reviewer and attached screenshots of the strings in the product

Asset Service - Asset route for the query assets availability report.