Merged PR 118231: Add support for query-ing assets that are currently being utilized
# Justification
Currently, there is no fixed API that tells the user what assets are currently being marked as being utilized by an utilization session. The user would need to use the "query-asset-utilization-history" route with some specific utilization filters to get the active utilization entries. After that, the user would have to use the "query-assets" route with some specific asset filters (based on the response from the "query-asset-utilization-history" route) to now get information about the assets currently being utilized.

The "Assets" grid needs to display asset summary information in terms of how many assets are currently being utilized. The UI will call into this new route to only retrieve the assets currently being marked for utilization

# Implementation
Add a new "query-utilized-assets" route that allows users to query for assets for which an utilization session is still in progress

The route has support for:
- asset filters -> Query for specific asset properties
- utilization filters -> Query for specific utilization properties

Note: The associated utilization sessions for assets to still be considered as "in use" have the following requirements:
- EndTimestamp should be NULL
- If there is NO HeartbeatTimestamp => StarTimestamp should be greater than DateTime.Now - 10 minutes
- If there is a HeartbeatTimestamp => HeartbeatTimestamp should be greater than DateTime.Now - 10 minutes
-> Basically the status of the utilization session should have been updated at least once in the last 10 minutes when the request was being made)

Added a stub route to the "AssetPerformanceManagement" HTTP Service because of Swagger

# Testing
- Added facade level unit tests
- Added integration tests

# Checklist
- [x] I tested changes to product code in product
- [ ] I considered updates to the wiki
- [ ] If this PR affects user-visible strings, I added Jenna Jaco as a reviewer and attached screenshots of the strings in the product